Fix dockerfile, rename schemas to filters
