+++ Dockerfile
@@ -6,7 +6,6 @@ ENV PYTHONUNBUFFERED 1
 RUN apk update && apk upgrade && \
   apk add --update bash python3 python3-dev postgresql-client postgresql-dev build-base gettext
 RUN mkdir /src
-ADD env /src
 ADD src /src
 WORKDIR /src
 RUN pip3 install --upgrade pip && pip3 install -r requirements.txt
